Split test_tag and testx_tag
